Added clear/reset button to the html
The "Clear" button in the HTML file is used to reset the input fields and clear the process table and Gantt chart. When the button is clicked, it triggers the clearAll() function, which should be defined in the script.js file. This function will handle the logic to clear all the relevant data and reset the visualizer to its initial state.